move tests into folders
