+++ test/collectAndCombine/LikeTermCollector.test.js
@@ -3,9 +3,9 @@
 const assert = require('assert');
 const math = require('mathjs');
 
-const LikeTermCollector = require('../lib/collectAndCombine/LikeTermCollector');
-const flatten = require('../lib/flattenOperands');
-const print = require('../lib/util/print');
+const LikeTermCollector = require('../../lib/collectAndCombine/LikeTermCollector');
+const flatten = require('../../lib/flattenOperands');
+const print = require('../../lib/util/print');
 
 function testCollectLikeTerms(exprStr, outputStr, explanation='', debug=false) {
   let description = `${exprStr} -> ${outputStr}`;
