+++ test/collectAndCombine/collectAndCombineSearch.test.js
@@ -3,9 +3,9 @@
 const assert = require('assert');
 const math = require('mathjs');
 
-const collectAndCombineSearch = require('../lib/collectAndCombine/collectAndCombineSearch');
-const flatten = require('../lib/flattenOperands');
-const print = require('../lib/util/print');
+const collectAndCombineSearch = require('../../lib/collectAndCombine/collectAndCombineSearch');
+const flatten = require('../../lib/flattenOperands');
+const print = require('../../lib/util/print');
 
 function testCollectAndCombineSubsteps(exprString, outputList, outputStr) {
   it(exprString + ' -> ' + outputStr, function () {
@@ -18,10 +18,11 @@ function testCollectAndCombineSubsteps(exprString, outputList, outputStr) {
         print(step.newNode),
         outputList[i]);
     });
-
-    assert.deepEqual(
-      print(status.newNode),
-      outputStr);
+    if (outputStr) {
+      assert.deepEqual(
+        print(status.newNode),
+        outputStr);
+    }
   });
 }
 
@@ -34,12 +35,17 @@ function testSimpleCollectAndCombine(exprString, outputStr) {
   });
 }
 
-describe('collectAndCombine with substeps', function () {
+describe('combinePolynomialTerms multiplication', function() {
   const tests = [
-    ['2x + 4x + y',
-      ['(2x + 4x) + y',
-        '6x + y'],
-      '6x + y'
+    ['x^2 * x * x',
+      ['x^2 * x^1 * x^1',
+        'x^(2 + 1 + 1)',
+        'x^4'],
+    ],
+    ['y * y^3',
+      ['y^1 * y^3',
+        'y^(1 + 3)',
+        'y^4'],
     ],
     ['2x * x^2 * 5x',
       ['(2 * 5) * (x * x^2 * x)',
@@ -51,7 +57,28 @@ describe('collectAndCombine with substeps', function () {
   tests.forEach(t => testCollectAndCombineSubsteps(t[0], t[1], t[2]));
 });
 
-describe('simple collectAndCombine with no substeps', function () {
+describe('combinePolynomialTerms addition', function() {
+  const tests = [
+    ['x+x',
+      ['1x + 1x',
+        '(1 + 1)x',
+        '2x']
+    ],
+    ['4y^2 + 7y^2 + y^2',
+      ['4y^2 + 7y^2 + 1y^2',
+        '(4 + 7 + 1)y^2',
+        '12y^2']
+    ],
+    ['2x + 4x + y',
+      ['(2x + 4x) + y',
+        '6x + y'],
+      '6x + y'
+    ],
+  ];
+  tests.forEach(t => testCollectAndCombineSubsteps(t[0], t[1]));
+});
+
+describe('collectAndCombine with no substeps', function () {
   const tests = [
     ['2x + 4x + x', '7x'],
     ['x * x^2 * x', 'x^4']
