+++ test/evaluateArithmeticSearch.test.js
